display: removed the code exhausts file descriptor in mdp overlay and rotator.
Fix the problem crashed while playing video clips and VT camera test.
ion_share_dma_buf is called in pr_debug.
So, it is called depends on build option and makes crash problem.

-------------
A. Dodd note:  Part of this patch was already in mdp4_overlay.c due to
the mako bulk updates patch.  This is the remaining part of the upstream
CAF patch.

Change-Id: I6355a7e9867d985cae229f1ef5316d2c5660abcd
Signed-off-by: Andrew Dodd <atd7@cornell.edu>