msm_rotator: Fast YUV logic improvements from Sony kernels
This plus the other patches improves landscape videos,
but portrait orientation videos still cause issues.

Starting to look like the longterm goal needs to be a CAF rebase,
or a whole pile of cherrypicking from upstream.  We've cherrypicked
nearly all the rotator patches we can.  Other possibility is
stuff from the Sony APQ8064 kernel that wasn't in the big mako
bulk update.

Change-Id: Id321ca8290d99e9b4789b4d5f4cd048f14029545