msm_fb: [HACK] Hardcode panel physical width/height
Some applications seem to be using the kernel/gralloc dpi settings,
which default to 160 if these are not set, instead of the system
dpi setting of 480.  This results in some applications displaying
TINY UI elements (such as Hacker's Keyboard).

Hardcode it until we can properly implement panel physical size
detection.

Change-Id: Ie5312cba1b59bead39bee2c69cce45a5be49481b